+++ sbpl_adaptive/include/sbpl_adaptive/sparse_adaptive_grid_3d.h
@@ -78,10 +78,10 @@ public:
 
     void world2grid(
         double wx, double wy, double wz,
-        size_t &gx, size_t &gy, size_t &gz) const;
+        int &gx, int &gy, int &gz) const;
 
     void grid2world(
-        size_t gx, size_t gy, size_t gz,
+        int gx, int gy, int gz,
         double& wx, double& wy, double& wz) const;
     ///@}
 
@@ -175,9 +175,9 @@ private:
 
     void addSphere(
         bool tracking,
-        size_t x,
-        size_t y,
-        size_t z,
+        int x,
+        int y,
+        int z,
         int rad,
         int near_rad,
         int dimID,
@@ -186,9 +186,9 @@ private:
 
     bool setCellNearDim(
         bool tracking,
-        size_t x,
-        size_t y,
-        size_t z,
+        int x,
+        int y,
+        int z,
         int dimID);
 
     void addTrackingSphere(
@@ -196,7 +196,7 @@ private:
         std::vector<Position3D> &modCells);
 
     void getOverlappingSpheres(
-        size_t x, size_t y, size_t z,
+        int x, int y, int z,
         int dimID,
         std::vector<std::vector<int>> &spheres);
 
@@ -446,7 +446,7 @@ void SparseAdaptiveGrid3D::addPlanningSphere(
     const AdaptiveSphere3D &sphere,
     std::vector<Position3D> &modCells)
 {
-    size_t gx, gy, gz;
+    int gx, gy, gz;
     world2grid(sphere.x,sphere.y,sphere.z,gx,gy,gz);
     int r = round(sphere.rad / oc_grid_->resolution());
     int nr = round(sphere.near_rad / oc_grid_->resolution());
@@ -475,17 +475,17 @@ const AdaptiveGridCell &SparseAdaptiveGrid3D::getCell(int gx, int gy, int gz) co
 inline
 const AdaptiveGridCell &SparseAdaptiveGrid3D::getCell(double wx, double wy, double wz) const
 {
-    size_t gcoordx, gcoordy, gcoordz;
+    int gcoordx, gcoordy, gcoordz;
     world2grid(wx, wy, wz, gcoordx, gcoordy, gcoordz);
-    return getCell((int)gcoordx, (int)gcoordy, (int)gcoordz);
+    return getCell(gcoordx, gcoordy, gcoordz);
 }
 
 inline
 unsigned int SparseAdaptiveGrid3D::getCellCostToGoal(double wx, double wy, double wz) const
 {
-    size_t gcoordx, gcoordy, gcoordz;
+    int gcoordx, gcoordy, gcoordz;
     world2grid(wx, wy, wz, gcoordx, gcoordy, gcoordz);
-    return getCellCostToGoal((int)gcoordx, (int)gcoordy, (int)gcoordz);
+    return getCellCostToGoal(gcoordx, gcoordy, gcoordz);
 }
 
 inline
@@ -497,9 +497,9 @@ bool SparseAdaptiveGrid3D::isInBounds(int gx, int gy, int gz) const
 inline
 bool SparseAdaptiveGrid3D::isInBounds(double wx, double wy, double wz) const
 {
-    size_t gx, gy, gz;
+    int gx, gy, gz;
     world2grid(wx, wy, wz, gx, gy, gz);
-    return isInBounds((int)gx, (int)gy, (int)gz);
+    return isInBounds(gx, gy, gz);
 }
 
 inline
@@ -507,7 +507,7 @@ void SparseAdaptiveGrid3D::addTrackingSphere(
     const AdaptiveSphere3D &sphere,
     std::vector<Position3D> &modCells)
 {
-    size_t gx, gy, gz;
+    int gx, gy, gz;
     world2grid(sphere.x, sphere.y, sphere.z, gx, gy, gz);
     int r = round(sphere.rad / oc_grid_->resolution());
     int nr = round(sphere.near_rad / oc_grid_->resolution());
